// LinkERP/ClientApp/src/app/models/inv/lbs-inv-inventory-stock-take-detail-product-style-matrix.ts
import { BaseEntity } from "../base-entity"
export class LBSINVInventoryStockTakeDetailProductStyleMatrix extends BaseEntity {
  CompanyID: any;
  StockTakeNo: any;
  ProductID: any;
  BinID: any;
  ProductMatrixRow: any;
  ProductMatrixColumn: any;
  SerialNo: any;
  SourceReference: any;
  TransactionDateIn: any;
  ExpiryDate: any;
  CurrentAvailableQuantity: any;
  CountQuantity: any;
  VarianceQuantity: any;
}